Refactor location
This commit change the file and line number to make them non optional.

This will prevent certain invalid location from existing, since, for debugger, knowing that a instruction is related to a file
 but not knowing were is not meaningful.

This commit also change the line and column number from u32 to NonZeroU32.

In dwarf line and column number start at 1.
if the line or column number were 0 it would mean that they are not specifies.

Sadly, those change break the testing apparatus of addr2line. As they verify if the create produce the same result as gnu addr2line.
 Which include incomplete location (i.e instruction with know file, but unknown line and column number).

I left fixing this problem as a todo