Change Location to a enum
before this commit, Location was a struct with 3 optional field : file, line and column.
This could lead to intense of Location are not pointing to a valid source location (eg file: None, line: some(..), column: Some(..)).

This commit remove to possibility of such invalid Location by transforming it into a enum.

This commit also add doc to the logic that determine if a set of file number, line number and column number form a valid location.

Since the change to a enum require to rework that section, I toke the opportunity to add more documentation.