Refactor set_breakpoint
This commit breakup the set_breakpoint into multiple function.

First, by separating the logic to get the address from the one which put the breakpoint.
And second, by putting the logic for parsing the address and the symbols into their own function.

Those change will help to add line to addr functionality in a later commit.