Add source location breakpoints to repl
this commit add the ability to set a  breakpoints by specifying the file name, the line number and (optional) the column number of the source code.

While it work, it is not very polish.

The error handling is sloppy.
It does no path canonization.
I have not update the UI the advertise the capability.
There are no path auto-completion.

In other word, it still is a prototype, but it work.